Adding optional config key for log max files
Problem:
Logs are removed after 5 days (5 files), and it's not configurable (or I did not found how). It was configurable in laravel 4, and I think it was a good idea (some people might want to keep 30 log files)

Solution : 
Add an optional config key `app.log_max_files` (with underscore as `locale` and `fallback_locale`). It will not break backward compatibility and we don't need to "hack" Laravel log system.  

Additional note:
Laravel's documentation should be updated if this PR is accepted.