[5.7] Queued Closures (#25777)
* Initial pass at queuable closures.

This reintroduces a feature that was previously present in early versions of Laravel queues. However, there have been improvements to serializable closure libraries in the meantime which allows for better security and better handling of Eloquent models and collections. Namely, signing closures with a hash to prevent modification and arbitrary code execution, as well as transforming and resolving models and collections to ModelIdentifier instances.

* Apply fixes from StyleCI (#25776)

* use config on app

* add dependency